% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_box_score_players.R
\docType{data}
\name{espn_game_id}
\alias{espn_game_id}
\title{Get box score data for the players, given a gameId}
\format{
An object of class \code{numeric} of length 1.
}
\usage{
espn_game_id
}
\arguments{
\item{espn_game_id}{(str) The eight digit gameId for the desired game}
}
\value{
A dataframe with player-level box-score information
}
\description{
Requires use of ESPN oriented team names, found in locally saved file (ids.csv)
}
\keyword{datasets}
