% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_box_score_players.R
\name{get_box_score_players}
\alias{get_box_score_players}
\title{Get box score data for the players, given a gameId}
\usage{
get_box_score_players(espn_game_id)
}
\description{
Requires use of ESPN oriented team names, found in locally saved file (ids.csv)
}
