% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_box_score_team.R
\name{get_box_score_team}
\alias{get_box_score_team}
\title{Get box score data for the teams, given a gameId}
\usage{
get_box_score_team(espn_game_id)
}
\arguments{
\item{espn_game_id}{(str) The eight digit gameId for the desired game}
}
\value{
A dataframe with team-level box-score information
}
\description{
Requires use of ESPN oriented team names, found in locally saved file (ids.csv)
}
