% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_game_data.R
\name{get_game_data}
\alias{get_game_data}
\title{Get game data (play-by-play, odds, and game details) for a given gameId}
\usage{
get_game_data(espn_game_id)
}
\arguments{
\item{espn_game_id}{(str) The eight digit gameId for the desired game}
}
\value{
A dataframe with complete game play-by-play data, along with season and game info
}
\description{
Requires use of ESPN oriented team names, found in locally saved file (ids.csv)
}
