% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_kenpom_data.R
\name{get_kenpom_data}
\alias{get_kenpom_data}
\title{Get season-level data for each D1 team from Ken Pomeroy's website}
\usage{
get_kenpom_data(season)
}
\arguments{
\item{season}{(str) The eight digit gameId for the desired game}
}
\value{
A dataframe with complete season data for kenpom
}
\description{
Get season-level data for each D1 team from Ken Pomeroy's website
}
