% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_odds_data.R
\name{get_odds_data}
\alias{get_odds_data}
\title{Get odds data for a given gameId}
\usage{
get_odds_data(espn_game_id)
}
\arguments{
\item{espn_game_id}{(str) The eight digit gameId for the desired game}
}
\value{
A dataframe with complete odds data, along with season and game info
}
\description{
Notes:
ESPN odds data only go back to 2012-13 season
Requires use of ESPN oriented team names, found in locally saved file (ids.csv)
}
