% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_season_schedule.R
\name{get_season_schedule}
\alias{get_season_schedule}
\title{Get a full master schedule of all games in a season for a list of teams}
\usage{
get_season_schedule(team, season)
}
\arguments{
\item{team}{(str) Team to get Play-by-Play data for}

\item{season}{(str) or (int) Season for which to get schedule. In form "2019-20" or "2020". Single number is season of tournament.}
}
\value{
A data frame of the team, gameIds, and season
}
\description{
Can use a single team, a list of teams in a conference, or full D1 list
Teams must be in ESPN format, found in ids file (ids.csv)
}
