% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ids.R
\docType{data}
\name{ids}
\alias{ids}
\title{Team ids required for pulling data from ESPN.com}
\format{
A data frame with 357 rows and 3 variables:
\describe{
  \item{team}{character team name on ESPN}
  \item{id}{numeric team id on ESPN}
  \link{link}{url link for ESPN website}
  ...
}
}
\source{
\url{http://www.espn.com/}
}
\usage{
ids
}
\description{
A dataset containing the team names and ids for all Division 1 NCAA basketball teams
}
\keyword{datasets}
