% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_bracket.R
\name{parse_bracket}
\alias{parse_bracket}
\title{Generates a dataframe of predictions ready to be printed to bracket}
\usage{
parse_bracket(prediction_file, year = 2019)
}
\description{
Given a csv file of kaggle game predictions, and season, this function
returns data in a format which can be easily plotted
}
\details{
Convert Kaggle predictions to bracket format
from zachmayer/kaggleNCAA on github
}
